requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.5
Django==6.0.3
invoke==2.2.1
mysql-connector-python==9.6.0
paramiko==4.0.0
pycparser==3.0
PyNaCl==1.6.2
RouterOS-api==0.21.0
sqlparse==0.5.5
tzdata==2025.3